Archtecture: Sequelize models were moved to database folder where should have been in first place since they exist to stablish connections between the database and the api
